Fix duplicate WebSocket notifications scaling with connected devices
A WebSocket client received N identical copies of every notification, where
N is the number of concurrently connected sockets/devices for the identity
(2 devices -> 2 copies, 3 devices -> 3 copies).

Root cause: AppNotificationHandler and PeerAppNotificationHandler were
registered InstancePerLifetimeScope, so each WebSocket connection (its own
request scope) got its own handler instance and therefore its own
RefCountedSubscription. EstablishConnection then registered one broker
subscription per connection on the shared pub/sub channel. The broker fans
each published message out to all subscriptions, and each subscription's
fan-out iterates the shared singleton DeviceSocketCollection and sends to
every socket. Net: 1 notification -> 1 publish -> N subscription callbacks
-> each loops N sockets -> every device receives N copies.

RefCountedSubscription was designed to be one shared, ref-counted
subscription per tenant, but living at per-connection scope defeated that.

Fix: extract the subscription + notification fan-out (which depend only on
singleton-safe services) into per-tenant singletons AppNotificationDispatcher
and PeerAppNotificationDispatcher. Each owns the single RefCountedSubscription,
so connection #1 registers the one broker subscription and the rest ref-count
onto it; the last to disconnect tears it down. The handlers stay scoped
(unchanged MediatR lifetime) and keep the per-connection command loop, which
needs request/DB-bound services (PeerInboxProcessor, PeerAppNotificationService);
they now delegate add-socket/subscribe/unsubscribe/remove-socket/send/publish
to the dispatcher. The in-flight inbox-drain coalescer also moves to the
singleton, so it now coalesces per drive per tenant rather than per connection.

Adds WebSocketDuplicateFanoutTests: opens two sockets on one identity, does a
single local upload, and asserts each socket receives the FileAdded exactly
once (failed with 2 before the fix, passes after).
