+++ src/services/Odin.Services/AppNotifications/WebSocket/AppNotificationHandler.cs
@@ -1,82 +1,55 @@
-﻿using System;
-using System.Collections.Concurrent;
+using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Net.WebSockets;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
-using Autofac;
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Odin.Core;
 using Odin.Core.Exceptions;
-using Odin.Core.Json;
 using Odin.Core.Serialization;
-using Odin.Core.Storage;
-using Odin.Core.Storage.PubSub;
 using Odin.Services.AppNotifications.ClientNotifications;
 using Odin.Services.Base;
 using Odin.Services.Drives;
-using Odin.Services.Drives.DriveCore.Storage;
-using Odin.Services.Drives.FileSystem.Base;
-using Odin.Services.Drives.Management;
 using Odin.Services.Mediator;
 using Odin.Services.Peer;
 using Odin.Services.Peer.Incoming.Drive.Transfer;
-using Odin.Services.Tenant.Container;
 
 #nullable enable
 
-// SEB:TODO this file needs to be combined with AppNotificationHandler.cs
-
-// SEB:TODO this file needs to be split up into different classes handling
-// - websocket incoming and outgoing messages
-// - client notifications (i.e. messages coming from the "inside" and going out to a websocket)
-
-// SEB:TODO cleanup SendMessageAsync params
-
 namespace Odin.Services.AppNotifications.WebSocket
 {
+    /// <summary>
+    /// Scoped MediatR handler and per-connection WebSocket command processor.
+    ///
+    /// The shared subscription and the fan-out of notifications to all connected sockets live
+    /// in the per-tenant singleton <see cref="AppNotificationDispatcher"/>; this class delegates
+    /// to it. Keeping this class scoped lets it use request/DB-bound services (e.g.
+    /// <see cref="PeerInboxProcessor"/>) for the per-connection command loop without making them
+    /// captive dependencies of a singleton.
+    /// </summary>
     public class AppNotificationHandler :
         INotificationHandler<IClientNotification>,
         INotificationHandler<IDriveNotification>,
         INotificationHandler<InboxItemReceivedNotification>
     {
-        private const string NotificationChannel = nameof(AppNotificationHandler);
-
         private readonly ILogger<AppNotificationHandler> _logger;
-        private readonly ITenantRootScope _tenantRootScope;
-        private readonly ITenantPubSub _pubSub;
+        private readonly AppNotificationDispatcher _dispatcher;
         private readonly PeerInboxProcessor _peerInboxProcessor;
-        private readonly SharedDeviceSocketCollection<AppNotificationHandler> _deviceSocketCollection;
-        private readonly RefCountedSubscription _notificationSubscription;
-
-        // Per-drive coalescer for server-side inbox drains triggered by InboxItemReceived
-        // notifications. The cached gate (TableInboxCached.GetReadyCountAsync) keeps repeat
-        // drains cheap, but under arrival bursts we'd still do redundant DB work; this keeps
-        // it to one in-flight drain per drive per process. Lazy<Task> ensures GetOrAdd's
-        // value factory races never start more than one drain task.
-        private readonly ConcurrentDictionary<Guid, Lazy<Task>> _inFlightDrains = new();
 
         public AppNotificationHandler(
             ILogger<AppNotificationHandler> logger,
-            ITenantRootScope tenantRootScope,
-            ITenantPubSub pubSub,
-            PeerInboxProcessor peerInboxProcessor,
-            SharedDeviceSocketCollection<AppNotificationHandler> deviceSocketCollection)
+            AppNotificationDispatcher dispatcher,
+            PeerInboxProcessor peerInboxProcessor)
         {
             _logger = logger;
-            _tenantRootScope = tenantRootScope;
-            _pubSub = pubSub;
+            _dispatcher = dispatcher;
             _peerInboxProcessor = peerInboxProcessor;
-            _deviceSocketCollection = deviceSocketCollection;
-            _notificationSubscription = new RefCountedSubscription(_pubSub, NotificationChannel, NotificationHandler);
         }
 
         //
-
 
         /// <summary>
         /// Awaits the configuration when establishing a new web socket connection
@@ -94,9 +67,9 @@ namespace Odin.Services.AppNotifications.WebSocket
                     Key = webSocketKey,
                     Socket = webSocket,
                 };
-                _deviceSocketCollection.AddSocket(deviceSocket);
+                _dispatcher.AddSocket(deviceSocket);
 
-                await _notificationSubscription.SubscribeAsync(cancellationToken);
+                await _dispatcher.SubscribeAsync(cancellationToken);
                 await AwaitCommands(deviceSocket, odinContext, cancellationToken);
             }
             catch (OperationCanceledException)
@@ -114,8 +87,8 @@ namespace Odin.Services.AppNotifications.WebSocket
             }
             finally
             {
-                await _notificationSubscription.UnsubscribeAsync(CancellationToken.None);
-                await _deviceSocketCollection.RemoveSocket(webSocketKey);
+                await _dispatcher.UnsubscribeAsync(CancellationToken.None);
+                await _dispatcher.RemoveSocket(webSocketKey);
                 _logger.LogTrace("WebSocket closed");
             }
         }
@@ -199,72 +172,14 @@ namespace Odin.Services.AppNotifications.WebSocket
         }
 
         //
-
-        private async Task NotificationHandler(JsonEnvelope envelope)
-        {
-            var message = envelope.DeserializeMessage();
-
-            switch (message)
-            {
-                case ClientNotificationMessage notification:
-                    await WsPublishAsync(notification);
-                    return;
-                case DriveNotificationMessage notification:
-                    await WsPublishAsync(notification);
-                    return;
-                case InboxItemReceivedNotificationMessage notification:
-                    await WsPublishAsync(notification);
-                    return;
-                case null:
-                    _logger.LogError("Received null message");
-                    return;
-                default:
-                    _logger.LogError("Unknown message type: {message}", message.GetType().Name);
-                    return;
-            }
-        }
-
-        //
         // IClientNotification
         //
 
         // Src: MediatR
-        // Dst: PubSub queue
-        public async Task Handle(IClientNotification notification, CancellationToken cancellationToken = default)
+        // Dst: PubSub queue (via the per-tenant dispatcher)
+        public Task Handle(IClientNotification notification, CancellationToken cancellationToken = default)
         {
-            var shouldEncrypt =
-                !(notification.NotificationType is ClientNotificationType.ConnectionRequestAccepted
-                    or ClientNotificationType.ConnectionRequestReceived);
-
-            var json = OdinSystemSerializer.Serialize(new
-            {
-                notification.NotificationType,
-                Data = notification.GetClientData()
-            });
-
-            var message = new ClientNotificationMessage
-            {
-                ShouldEncrypt = shouldEncrypt,
-                Json = json,
-            };
-
-            await _pubSub.PublishAsync(NotificationChannel, JsonEnvelope.Create(message));
-        }
-
-        //
-
-        // Src: PubSub queue
-        // Dst: WebSocket
-        private async Task WsPublishAsync(ClientNotificationMessage notification)
-        {
-            var sockets = _deviceSocketCollection.GetAll().Values;
-            foreach (var deviceSocket in sockets)
-            {
-                await SendMessageAsync(
-                    deviceSocket,
-                    notification.Json,
-                    notification.ShouldEncrypt);
-            }
+            return _dispatcher.PublishClientNotificationAsync(notification);
         }
 
         //
@@ -272,281 +187,39 @@ namespace Odin.Services.AppNotifications.WebSocket
         //
 
         // Src: MediatR
-        // Dst: PubSub queue
-        public async Task Handle(IDriveNotification notification, CancellationToken cancellationToken = default)
+        // Dst: PubSub queue (via the per-tenant dispatcher)
+        public Task Handle(IDriveNotification notification, CancellationToken cancellationToken = default)
         {
-            var message = new DriveNotificationMessage
-            {
-                NotificationType = notification.NotificationType,
-                File = notification.File,
-                IsDeleteNotification = notification is DriveFileDeletedNotification,
-                ServerFileHeader = notification.ServerFileHeader,
-                PreviousServerFileHeader = (notification as DriveFileDeletedNotification)?.PreviousServerFileHeader
-            };
-
-            await _pubSub.PublishAsync(NotificationChannel, JsonEnvelope.Create(message));
-        }
-
-        //
-
-        // Src: PubSub queue
-        // Dst: WebSocket
-        private async Task WsPublishAsync(DriveNotificationMessage notification)
-        {
-            var sockets = _deviceSocketCollection.GetAll().Values
-                .Where(ds => ds.Drives.Any(driveId => driveId == notification.File.DriveId))
-                .ToList();
-
-            if (sockets.Count == 0)
-            {
-                return;
-            }
-
-            await using var scope = _tenantRootScope.BeginLifetimeScope();
-            if (scope == null)
-            {
-                return;
-            }
-
-            var driveManager =  scope.Resolve<IDriveManager>();
-            var drive = await driveManager.GetDriveAsync(notification.File.DriveId);
-
-            foreach (var deviceSocket in sockets)
-            {
-                var deviceOdinContext = deviceSocket.DeviceOdinContext;
-                var hasSharedSecret = null != deviceOdinContext?.PermissionsContext?.SharedSecretKey;
-
-                var o = new ClientDriveNotification
-                {
-                    TargetDrive = drive?.TargetDriveInfo,
-                    Header = hasSharedSecret
-                        ? DriveFileUtility.CreateClientFileHeader(notification.ServerFileHeader, deviceOdinContext)
-                        : null,
-                    PreviousServerFileHeader = hasSharedSecret
-                        ? notification.IsDeleteNotification
-                            ? DriveFileUtility.CreateClientFileHeader(notification.PreviousServerFileHeader, deviceOdinContext!)
-                            : null
-                        : null
-                };
-
-                var json = OdinSystemSerializer.Serialize(new
-                {
-                    notification.NotificationType,
-                    Data = OdinSystemSerializer.Serialize(o)
-                });
-
-                await SendMessageAsync(deviceSocket, json, encrypt: true);
-            }
+            return _dispatcher.PublishDriveNotificationAsync(notification);
         }
 
         //
         // InboxItemReceivedNotification
         //
 
-        // TODO: This notification is largely redundant now. TryDrainInboxForDriveAsync runs
-        // synchronously before we send InboxItemReceived to clients, and any
-        // DriveFileAddedNotification it produces fans out as FileAdded over the same WS
-        // (via Handle(IDriveNotification)) — carrying the actual file header. The bare
-        // "something arrived" ping is therefore useful only when no FileAdded will follow.
-        // Suppress per TransferFileType once clients have migrated: keep it for types that
-        // don't produce a FileAdded (e.g. read receipts, reactions) and drop it for the rest.
-        // Note: empirical testing showed InboxItemReceived does not currently reach the
-        // owner WS in the peer-file-arrival scenario at all — there is little owner-WS
-        // back-compat surface to preserve. The peer-app WS path (PeerAppNotificationHandler)
-        // may differ; verify before removing.
-
         // Src: MediatR
-        // Dst: PubSub queue
-        public async Task Handle(InboxItemReceivedNotification notification, CancellationToken cancellationToken = default)
+        // Dst: PubSub queue (via the per-tenant dispatcher)
+        public Task Handle(InboxItemReceivedNotification notification, CancellationToken cancellationToken = default)
         {
-            var message = new InboxItemReceivedNotificationMessage
-            {
-                NotificationType = notification.NotificationType,
-                TargetDrive = notification.TargetDrive,
-                FileSystemType = notification.FileSystemType,
-                TransferFileType = notification.TransferFileType,
-            };
-
-            await _pubSub.PublishAsync(NotificationChannel, JsonEnvelope.Create(message));
-        }
-
-        // Src: PubSub queue
-        // Dst: WebSocket
-        private async Task WsPublishAsync(InboxItemReceivedNotificationMessage notification)
-        {
-            var notificationDriveId = notification.TargetDrive.Alias;
-
-            // Opportunistic server-side inbox drain, mirroring the QueryBatch path.
-            // We use an owner-acting context already cached on a locally-subscribed
-            // DeviceSocket; any DriveFileAdded events the drain produces will fan out
-            // through Handle(IDriveNotification) onto the same sockets.
-            await TryDrainInboxForDriveAsync(notificationDriveId);
-
-            var translated = new TranslatedClientNotification(notification.NotificationType,
-                OdinSystemSerializer.Serialize(new
-                {
-                    TargetDrive = notification.TargetDrive,
-                    notification.TransferFileType,
-                    notification.FileSystemType
-                }));
-
-            await SerializeSendToAllDevicesForDrive(notificationDriveId, translated, false);
+            return _dispatcher.PublishInboxItemReceivedAsync(notification);
         }
 
         //
 
-        private async Task TryDrainInboxForDriveAsync(Guid driveId)
+        private Task SendErrorMessageAsync(DeviceSocket deviceSocket, string errorText, CancellationToken cancellationToken = default)
         {
-            var deviceOdinContext = _deviceSocketCollection.GetAll().Values
-                .Select(ds => ds.DeviceOdinContext)
-                .FirstOrDefault(ctx => ctx != null && PeerInboxDriveQueue.IsAuthorizedToDrain(driveId, ctx));
-
-            if (deviceOdinContext == null)
-            {
-                return;
-            }
-
-            var lazy = _inFlightDrains.GetOrAdd(driveId,
-                id => new Lazy<Task>(() => RunDrainAsync(id, deviceOdinContext)));
-
-            try
-            {
-                await lazy.Value;
-            }
-            catch (Exception e)
-            {
-                // InboxDrainOnQuery already swallows and logs internally; this is defensive
-                // so a drain failure can never break the WS notification fan-out.
-                _logger.LogInformation(e, "Inbox drain on WS notify failed: {error}", e.Message);
-            }
-        }
-
-        private async Task RunDrainAsync(Guid driveId, IOdinContext deviceOdinContext)
-        {
-            try
-            {
-                await using var scope = _tenantRootScope.BeginLifetimeScope();
-                if (scope == null)
-                {
-                    return;
-                }
-
-                var drain = scope.Resolve<InboxDrainOnQuery>();
-                await drain.DrainIfReadyAsync(driveId, deviceOdinContext);
-            }
-            finally
-            {
-                _inFlightDrains.TryRemove(driveId, out _);
-            }
+            return _dispatcher.SendErrorMessageAsync(deviceSocket, errorText, cancellationToken);
         }
 
         //
 
-        private async Task SerializeSendToAllDevicesForDrive(
-            Guid targetDriveId,
-            IClientNotification notification,
-            bool encrypt,
-            CancellationToken cancellationToken = default)
-        {
-            var json = OdinSystemSerializer.Serialize(new
-            {
-                notification.NotificationType,
-                Data = notification.GetClientData()
-            });
-
-            var sockets = _deviceSocketCollection.GetAll().Values
-                .Where(ds => ds.Drives.Any(driveId => driveId == targetDriveId));
-
-            foreach (var deviceSocket in sockets)
-            {
-                if (!cancellationToken.IsCancellationRequested)
-                {
-                    await SendMessageAsync(deviceSocket, json, encrypt, cancellationToken);
-                }
-            }
-        }
-
-        //
-
-        private static string JsonMessage(ClientNotificationType notificationType, string message)
-        {
-            return OdinSystemSerializer.Serialize(new
-            {
-                NotificationType = notificationType,
-                Data = message,
-            });
-        }
-
-        //
-
-        private async Task SendErrorMessageAsync(DeviceSocket deviceSocket, string errorText, CancellationToken cancellationToken = default)
-        {
-            var json = JsonMessage(ClientNotificationType.Error, errorText);
-            await SendMessageAsync(
-                deviceSocket,
-                json,
-                deviceSocket.DeviceOdinContext?.PermissionsContext?.SharedSecretKey != null,
-                cancellationToken);
-        }
-
-        //
-
-        private async Task SendMessageAsync(
+        private Task SendMessageAsync(
             DeviceSocket deviceSocket,
             string message,
             bool encrypt,
             CancellationToken cancellationToken = default)
         {
-            var socket = deviceSocket.Socket;
-
-            if (socket is not { State: WebSocketState.Open } || cancellationToken.IsCancellationRequested)
-            {
-                return;
-            }
-
-            try
-            {
-                if (encrypt)
-                {
-                    if (deviceSocket.DeviceOdinContext == null)
-                    {
-                        await _deviceSocketCollection.RemoveSocket(deviceSocket.Key);
-                        _logger.LogInformation("Invalid/Stale Device found; removing from list; closing socket");
-                        throw new WebSocketException("Missing device odin context");
-                    }
-
-                    if (deviceSocket.DeviceOdinContext.PermissionsContext?.SharedSecretKey == null)
-                    {
-                        throw new OdinSystemException("Cannot encrypt message without shared secret key");
-                    }
-
-                    var key = deviceSocket.DeviceOdinContext.PermissionsContext.SharedSecretKey;
-                    var encryptedPayload = SharedSecretEncryptedPayload.Encrypt(message.ToUtf8ByteArray(), key);
-                    message = OdinSystemSerializer.Serialize(encryptedPayload);
-                }
-
-                var payload = new ClientNotificationPayload()
-                {
-                    IsEncrypted = encrypt,
-                    Payload = message
-                };
-
-                var json = OdinSystemSerializer.Serialize(payload);
-                await deviceSocket.FireAndForgetAsync(json, cancellationToken);
-            }
-            catch (OperationCanceledException)
-            {
-                // ignore, this exception is expected when the socket is closing behind the scenes
-            }
-            catch (WebSocketException e)
-            {
-                _logger.LogInformation("WebSocketException: {error}", e.Message);
-            }
-            catch (Exception e)
-            {
-                //HACK: need to find out what is trying to write when the response is complete
-                _logger.LogError(e, "SendMessageAsync: {error}", e.Message);
-            }
+            return _dispatcher.SendMessageAsync(deviceSocket, message, encrypt, cancellationToken);
         }
 
         //
@@ -629,27 +302,13 @@ namespace Odin.Services.AppNotifications.WebSocket
 
         //
 
-        private class ClientNotificationMessage
+        private static string JsonMessage(ClientNotificationType notificationType, string message)
         {
-            public required bool ShouldEncrypt { get; init; }
-            public required string Json { get; init; }
-        }
-
-        private class DriveNotificationMessage
-        {
-            public required ClientNotificationType NotificationType { get; init; }
-            public required InternalDriveFileId File { get; init; }
-            public required bool IsDeleteNotification { get; init; }
-            public required ServerFileHeader ServerFileHeader { get; init; }
-            public required ServerFileHeader? PreviousServerFileHeader { get; init; }
-        }
-
-        private class InboxItemReceivedNotificationMessage
-        {
-            public required ClientNotificationType NotificationType { get; init; }
-            public required TargetDrive TargetDrive { get; init; }
-            public required FileSystemType FileSystemType { get; init; }
-            public required TransferFileType TransferFileType { get; init; }
+            return OdinSystemSerializer.Serialize(new
+            {
+                NotificationType = notificationType,
+                Data = message,
+            });
         }
     }
 }
